+++ mountpoint-s3-client/src/s3_crt_client.rs
@@ -26,7 +26,6 @@ use mountpoint_s3_crt::io::channel_bootstrap::{ClientBootstrap, ClientBootstrapO
 pub use mountpoint_s3_crt::io::event_loop::EventLoopGroup;
 use mountpoint_s3_crt::io::host_resolver::{AddressKinds, HostResolver, HostResolverDefaultOptions};
 use mountpoint_s3_crt::io::retry_strategy::{ExponentialBackoffJitterMode, RetryStrategy, StandardRetryOptions};
-use mountpoint_s3_crt::io::stream::InputStream;
 use mountpoint_s3_crt::io::tls::{TlsConnectionOptions, TlsContext, TlsContextOptions, TlsError as CrtTlsError};
 use mountpoint_s3_crt::s3::buffer::Buffer;
 use mountpoint_s3_crt::s3::client::{
@@ -611,7 +610,7 @@ impl S3CrtClientInner {
     /// Pre-populates common headers used across all requests. Sets the "accept" header assuming the
     /// response should be XML; this header should be overwritten for requests like GET that return
     /// object data.
-    fn new_request_template(&self, method: &str, bucket: &str) -> Result<S3Message<'_>, ConstructionError> {
+    fn new_request_template(&self, method: &str, bucket: &str) -> Result<S3Message, ConstructionError> {
         let endpoint = self.endpoint_config.resolve_for_bucket(bucket)?;
         let uri = endpoint.uri()?;
         trace!(?uri, "resolved endpoint");
@@ -1056,8 +1055,8 @@ impl S3Operation {
 /// virtual-hosted-style addresses. The `path_prefix` is appended to the front of all paths, and
 /// need not be terminated with a `/`.
 #[derive(Debug)]
-struct S3Message<'a> {
-    inner: Message<'a>,
+struct S3Message {
+    inner: Message,
     uri: Uri,
     path_prefix: String,
     checksum_config: Option<ChecksumConfig>,
@@ -1119,7 +1118,7 @@ impl<P: AsRef<OsStr>> QueryFragment<'_, P> {
     }
 }
 
-impl<'a> S3Message<'a> {
+impl S3Message {
     /// Add a header to this message. The header is added if necessary and any existing values for
     /// this header are removed.
     fn set_header(
@@ -1159,12 +1158,6 @@ impl<'a> S3Message<'a> {
         self.checksum_config = checksum_config;
     }
 
-    /// Sets the body input stream for this message, and returns any previously set input stream.
-    /// If input_stream is None, unsets the body.
-    fn set_body_stream(&mut self, input_stream: Option<InputStream<'a>>) -> Option<InputStream<'a>> {
-        self.inner.set_body_stream(input_stream)
-    }
-
     /// Set the content length header.
     fn set_content_length_header(
         &mut self,
@@ -1189,7 +1182,7 @@ impl<'a> S3Message<'a> {
         self.inner.set_header(&header)
     }
 
-    fn into_options(self, operation: S3Operation) -> MetaRequestOptions<'a> {
+    fn into_options(self, operation: S3Operation) -> MetaRequestOptions {
         let mut options = MetaRequestOptions::new();
         if let Some(checksum_config) = self.checksum_config {
             options.checksum_config(checksum_config);
@@ -1670,12 +1663,12 @@ impl ObjectClient for S3CrtClient {
         self.put_object(bucket, key, params).await
     }
 
-    async fn put_object_single<'a>(
+    async fn put_object_single(
         &self,
         bucket: &str,
         key: &str,
         params: &PutObjectSingleParams,
-        contents: impl AsRef<[u8]> + Send + 'a,
+        contents: impl AsRef<[u8]> + Send + 'static,
     ) -> ObjectClientResult<PutObjectResult, PutObjectError, Self::ClientError> {
         self.put_object_single(bucket, key, params, contents).await
     }