+++ mountpoint-s3-crt/src/io.rs
@@ -11,7 +11,6 @@ pub mod event_loop;
 pub mod futures;
 pub mod host_resolver;
 pub mod retry_strategy;
-pub mod stream;
 pub mod tls;
 
 static IO_LIBRARY_INIT: Once = Once::new();